Update Write-ObjectToSQL.psm1
I had an issue where the first row of my $InputObject had some values that were from the DBNull type, and thus were ignored. By performing this update I was able to ignore those and use data from other rows as examples to the script, so that it could identify the correct Type for the column.